requirements/dev.txt
# Meta
-e .
-r ./base.txt

# Sessions
nox==2022.1.7

# Docs
furo
sphinx>=5,<6

# Formatting
black[jupyter]~=22.6.0

# Imports
flake8~=4.0.1
isort~=5.10.0

# Typing
mypy==0.961
hikari-lightbulb
hikari-crescent
types-attrs
types-setuptools

# Line lengths
len8~=0.7.2

# Spelling
codespell~=2.1.0

# Tests
coverage~=6.4.1
mock~=4.0.3
pytest~=7.1.0
pytest-asyncio~=0.18.1
requests>=2.27,<3

# Safety
safety>=2.0.0,<3

# Security
bandit~=1.7.2

# Dependencies
deputil~=0.2.0
